refactor jumbotrons.py
